--- backend/src/api/auth.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const authCtrl = require('./auth/register');
-const router = express.Router();
-
-/* All routes in authentication module */
-router.route('/register').post(authCtrl.apiPostTask);
-
-module.exports = router;
