+++ backend/src/models/user.js
@@ -18,4 +18,4 @@ const userSchema = new mongoose.Schema({
     }
 }, { timestamps: true});
 
-module.exports = mongoose.model('User', userSchema);
+module.exports = mongoose.model('User', userSchema); //Export data formatting
