+++ backend/src/route/register.route.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const registerCtrl = require('../api/register'); //Used to send data to the register endpoint
+const router = express.Router();
+/* All routes in authentication module */
+router.route('/register').post(registerCtrl.apiPostTask);
+
+module.exports = router; //Exports router for server
